+++ MoviePickerDomain/Model/Genre.cs
@@ -5,7 +5,7 @@ namespace MoviePickerDomain.Model;
 
 public partial class Genre : Entity
 {
-    public short GenreId { get; set; }
+    public long GenreId { get; set; }
 
     public string Name { get; set; } = null!;
 }
