+++ MoviePickerDomain/Model/MoviesGenre.cs
@@ -9,7 +9,7 @@ public partial class MoviesGenre : Entity
 
     public long MovieId { get; set; }
 
-    public short GenreId { get; set; }
+    public long GenreId { get; set; }
 
     public virtual Genre Genre { get; set; } = null!;
